bb84-service/requirements.txt
fastapi
uvicorn[standard]
pycryptodome
pydanticcd